eliminate circular dependency xrepo <-> commits, util <-> ...
