+++ lsif/src/util.ts
@@ -1,8 +1,8 @@
 import * as fs from 'mz/fs'
 import * as path from 'path'
-import { DefinitionReferenceResultId } from './database.models'
 import { Id } from 'lsif-protocol'
 import * as constants from './constants'
+import { DefinitionReferenceResultId } from './database.types'
 
 /**
  * Reads an integer from an environment variable or defaults to the given value.
